requirements.txt
Flask==3.1.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
playwright==1.52.0
requests==2.32.3
python-dotenv==1.1.0

